fix: remove deprecated force_upgrade argument (#637)
##### SUMMARY

The alias removal was planned for the version 5.0.0, but I forgot to
remove the alias before cutting the 5.0.0 release (the sanity check only
failed after the release pull request was merged).

Since this is the second time I forgot to remove the deprecated alias, I
will cut a patch release, including this breaking change, as soon as
possible, so we still consider this part of the 5.0.0 release.

##### ISSUE TYPE

- Bugfix Pull Request


##### COMPONENT NAME

server